start quarantining demo code, have outbuffer built into SimpleAudioSystem (managed IO)
