+++ src/main/scala/buffix/ugen/Phasor.scala
@@ -6,8 +6,7 @@ import buffix.AudioConfig
   * Created by johnmcgill on 3/5/18.
   */
 object Phasor {
-  val TwoPi = Math.PI * 2
-  val Period = TwoPi
+  val Period = Math.PI * 2
 
   def computeW(pitch: Int, samplingRate: Int): Double = Period * pitch / samplingRate
 }
